+++ datumaro/plugins/icdar_format/converter.py
@@ -10,73 +10,80 @@ from datumaro.components.extractor import AnnotationType, CompiledMask
 from datumaro.util.image import save_image
 from datumaro.util.mask_tools import paint_mask
 
-from .format import IcdarPath, IcdarTask
+from .format import IcdarPath
 
 
-class _WordRecognitionConverter:
-    def __init__(self):
-        self.annotations = ''
+class IcdarWordRecognitionConverter(Converter):
+    DEFAULT_IMAGE_EXT = IcdarPath.IMAGE_EXT
 
-    def save_annotations(self, item, path):
-        self.annotations += '%s, ' % (item.id + IcdarPath.IMAGE_EXT)
-        for ann in item.annotations:
-            if ann.type != AnnotationType.caption:
-                continue
-            self.annotations += '\"%s\"' % ann.caption
-        self.annotations += '\n'
+    def apply(self):
+        for subset_name, subset in self._extractor.subsets().items():
+            annotation = ''
+            for item in subset:
+                image_filename = self._make_image_filename(item)
+                if self._save_images and item.has_image:
+                    self._save_image(item, osp.join(self._save_dir,
+                        subset_name, IcdarPath.IMAGES_DIR, image_filename))
 
-    def write(self, path):
-        file = osp.join(path, 'gt.txt')
-        os.makedirs(osp.dirname(file), exist_ok=True)
-        with open(file, 'w') as f:
-            f.write(self.annotations)
+                annotation += '%s, ' % image_filename
+                for ann in item.annotations:
+                    if ann.type != AnnotationType.caption:
+                        continue
+                    annotation += '\"%s\"' % ann.caption
+                annotation += '\n'
 
-    def is_empty(self):
-        return len(self.annotations) == 0
+            if len(annotation):
+                anno_file = osp.join(self._save_dir, subset_name, 'gt.txt')
+                os.makedirs(osp.dirname(anno_file), exist_ok=True)
+                with open(anno_file, 'w', encoding='utf-8') as f:
+                    f.write(annotation)
 
-class _TextLocalizationConverter:
-    def __init__(self):
-        self.annotations = {}
+class IcdarTextLocalizationConverter(Converter):
+    DEFAULT_IMAGE_EXT = IcdarPath.IMAGE_EXT
 
-    def save_annotations(self, item, path):
-        annotation = ''
-        for ann in item.annotations:
-            if ann.type == AnnotationType.bbox:
-                annotation += ' '.join(str(p) for p in ann.points)
-                if ann.attributes and 'text' in ann.attributes:
-                    annotation += ' \"%s\"' % ann.attributes['text']
-            elif ann.type == AnnotationType.polygon:
-                annotation += ','.join(str(p) for p in ann.points)
-                if ann.attributes and 'text' in ann.attributes:
-                    annotation += ',\"%s\"' % ann.attributes['text']
-            annotation += '\n'
-        self.annotations[item.id] = annotation
+    def apply(self):
+        for subset_name, subset in self._extractor.subsets().items():
+            for item in subset:
+                if self._save_images and item.has_image:
+                    self._save_image(item,
+                        subdir=osp.join(subset_name, IcdarPath.IMAGES_DIR))
 
-    def write(self, path):
-        os.makedirs(path, exist_ok=True)
-        for item in self.annotations:
-            file = osp.join(path, 'gt_' + item + '.txt')
-            with open(file, 'w') as f:
-                f.write(self.annotations[item])
+                annotation = ''
+                for ann in item.annotations:
+                    if ann.type == AnnotationType.bbox:
+                        annotation += ' '.join(str(p) for p in ann.points)
+                        if ann.attributes and 'text' in ann.attributes:
+                            annotation += ' \"%s\"' % ann.attributes['text']
+                    elif ann.type == AnnotationType.polygon:
+                        annotation += ','.join(str(p) for p in ann.points)
+                        if ann.attributes and 'text' in ann.attributes:
+                            annotation += ',\"%s\"' % ann.attributes['text']
+                    annotation += '\n'
 
-    def is_empty(self):
-        return len(self.annotations) == 0
+                anno_file = osp.join(self._save_dir, subset_name,
+                    osp.dirname(item.id), 'gt_' + osp.basename(item.id) + '.txt')
+                os.makedirs(osp.dirname(anno_file), exist_ok=True)
+                with open(anno_file, 'w', encoding='utf-8') as f:
+                    f.write(annotation)
 
-class _TextSegmentationConverter:
-    def __init__(self):
-        self.annotations = {}
+class IcdarTextSegmentationConverter(Converter):
+    DEFAULT_IMAGE_EXT = IcdarPath.IMAGE_EXT
 
-    def save_annotations(self, item, path):
+    def apply(self):
+        for subset_name, subset in self._extractor.subsets().items():
+            for item in subset:
+                self._save_item(subset_name, subset, item)
+
+    def _save_item(self, subset_name, subset, item):
+        if self._save_images and item.has_image:
+            self._save_image(item,
+                subdir=osp.join(subset_name, IcdarPath.IMAGES_DIR))
+
         annotation = ''
         colormap = [(255, 255, 255)]
-        anns = [a for a in item.annotations
-            if a.type == AnnotationType.mask]
+        anns = [a for a in item.annotations if a.type == AnnotationType.mask]
         if anns:
-            is_not_index = len([p for p in anns if 'index' not in p.attributes])
-            if is_not_index:
-                raise Exception("Item %s: a mask must have"
-                    "'index' attribute" % item.id)
-            anns = sorted(anns, key=lambda a: a.attributes['index'])
+            anns = sorted(anns, key=lambda a: int(a.attributes.get('index', 0)))
             group = anns[0].group
             for ann in anns:
                 if ann.group != group or (not ann.group and anns[0].group != 0):
@@ -95,7 +102,7 @@ class _TextSegmentationConverter:
                         annotation += ' '.join(p for p in color)
                     else:
                         raise Exception("Item %s: a mask must have "
-                            "an RGB color attribute, e. g. '10 7 50'" % item.id)
+                            "an RGB color attribute, e.g. '10 7 50'" % item.id)
                     if 'center' in ann.attributes:
                         annotation += ' %s' % ann.attributes['center']
                     else:
@@ -111,84 +118,11 @@ class _TextSegmentationConverter:
                 instance_labels=[m.attributes['index'] + 1 for m in anns])
             mask = paint_mask(mask.class_mask,
                 { i: colormap[i] for i in range(len(colormap)) })
-            save_image(osp.join(path, item.id + '_GT' + IcdarPath.GT_EXT),
-                mask, create_dir=True)
-        self.annotations[item.id] = annotation
+            save_image(osp.join(self._save_dir, subset_name,
+                item.id + '_GT' + IcdarPath.GT_EXT), mask, create_dir=True)
 
-    def write(self, path):
-        os.makedirs(path, exist_ok=True)
-        for item in self.annotations:
-            file = osp.join(path, item + '_GT' + '.txt')
-            with open(file, 'w') as f:
-                f.write(self.annotations[item])
-
-    def is_empty(self):
-        return len(self.annotations) == 0
-
-
-class IcdarConverter(Converter):
-    DEFAULT_IMAGE_EXT = IcdarPath.IMAGE_EXT
-
-    _TASK_CONVERTER = {
-        IcdarTask.word_recognition: _WordRecognitionConverter,
-        IcdarTask.text_localization: _TextLocalizationConverter,
-        IcdarTask.text_segmentation: _TextSegmentationConverter,
-    }
-
-    def __init__(self, extractor, save_dir, tasks=None, **kwargs):
-        super().__init__(extractor, save_dir, **kwargs)
-
-        assert tasks is None or isinstance(tasks, (IcdarTask, list, str))
-        if isinstance(tasks, IcdarTask):
-            tasks = [tasks]
-        elif isinstance(tasks, str):
-            tasks = [IcdarTask[tasks]]
-        elif tasks:
-            for i, t in enumerate(tasks):
-                if isinstance(t, str):
-                    tasks[i] = IcdarTask[t]
-                else:
-                    assert t in IcdarTask, t
-        self._tasks = tasks
-
-    def _make_task_converter(self, task):
-        if task not in self._TASK_CONVERTER:
-            raise NotImplementedError()
-        return self._TASK_CONVERTER[task]()
-
-    def _make_task_converters(self):
-        return { task: self._make_task_converter(task)
-            for task in (self._tasks or self._TASK_CONVERTER) }
-
-    def apply(self):
-        for subset_name, subset in self._extractor.subsets().items():
-            task_converters = self._make_task_converters()
-            for item in subset:
-                for task, task_conv in task_converters.items():
-                    if item.has_image and self._save_images:
-                        self._save_image(item, osp.join(
-                            self._save_dir, subset_name, IcdarPath.IMAGES_DIR,
-                            item.id + IcdarPath.IMAGE_EXT))
-                    task_conv.save_annotations(item, osp.join(self._save_dir,
-                        IcdarPath.TASK_DIR[task], subset_name))
-
-            for task, task_conv in task_converters.items():
-                if task_conv.is_empty() and not self._tasks:
-                    continue
-                task_conv.write(osp.join(self._save_dir,
-                    IcdarPath.TASK_DIR[task], subset_name))
-
-class IcdarWordRecognitionConverter(IcdarConverter):
-    def __init__(self, *args, **kwargs):
-        kwargs['tasks'] = IcdarTask.word_recognition
-        super().__init__(*args, **kwargs)
-
-class IcdarTextLocalizationConverter(IcdarConverter):
-    def __init__(self, *args, **kwargs):
-        kwargs['tasks'] = IcdarTask.text_localization
-        super().__init__(*args, **kwargs)
-
-class IcdarTextSegmentationConverter(IcdarConverter):
-    def __init__(self, *args, **kwargs):
-        kwargs['tasks'] = IcdarTask.text_segmentation
-        super().__init__(*args, **kwargs)
+        anno_file = osp.join(self._save_dir, subset_name,
+            item.id + '_GT' + '.txt')
+        os.makedirs(osp.dirname(anno_file), exist_ok=True)
+        with open(anno_file, 'w', encoding='utf-8') as f:
+            f.write(annotation)